Move logic to get release notes from a RPM to a new class
